control-plane/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
pyyaml==6.0.1
httpx==0.27.0
